CC-38979 Pull validateSignalDataCollection template into JdbcConnection
Move the flag/config guard, try/catch, "not found" error, and column-shape
delegation onto JdbcConnection.validateSignalDataCollection, exposing just
the DB-specific identifier resolution as a protected hook
(resolveSignalDataCollectionTableId). PostgresConnection, SqlServerConnection,
and OracleConnection now inherit the base implementation unchanged - the
default schema-first parse + readTableNames probe matches what each of them
was doing. BinlogConnectorConnection keeps a small override for catalog-first
parsing and conditional TableId.toLowercase() per isTableIdCaseSensitive().

Net effect: ~100 lines of duplicated scaffolding removed across four
*Connection subclasses. All 31 existing unit + IT tests still green.
