+++ debezium-connector-binlog/src/main/java/io/debezium/connector/binlog/jdbc/BinlogConnectorConnection.java
@@ -30,7 +30,6 @@ import io.debezium.jdbc.JdbcConnection;
 import io.debezium.pipeline.spi.OffsetContext;
 import io.debezium.pipeline.spi.Partition;
 import io.debezium.relational.Column;
-import io.debezium.relational.SignalDataCollectionChecks;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
 import io.debezium.util.Strings;
@@ -192,40 +191,20 @@ public abstract class BinlogConnectorConnection extends JdbcConnection {
     }
 
     /**
-     * Validates the configured signal data collection (when validation is enabled) against
-     * MySQL/MariaDB semantics: a two-part {@code database.table} identifier where {@code database}
-     * is the catalog. When the server is case-insensitive (see {@link #isTableIdCaseSensitive()})
-     * the identifier is normalised to lowercase once — matching how the rest of the binlog
-     * connector treats table identifiers.
+     * MySQL/MariaDB parses {@code signal.data.collection} as {@code database.table} (catalog-first)
+     * and normalises the identifier to lowercase when the server is case-insensitive (see
+     * {@link #isTableIdCaseSensitive()}), matching how the rest of the binlog connector treats
+     * table identifiers.
      */
-    public List<String> validateSignalDataCollection(CommonConnectorConfig config) {
-        if (!config.isSignalDataCollectionValidationEnabled()) {
-            return Collections.emptyList();
+    @Override
+    protected TableId resolveSignalDataCollectionTableId(String raw) throws SQLException {
+        TableId parsed = TableId.parse(raw, true);
+        if (!isTableIdCaseSensitive()) {
+            parsed = parsed.toLowercase();
         }
-        final String raw = config.getSignalingDataCollectionId();
-        if (Strings.isNullOrBlank(raw)) {
-            return Collections.emptyList();
-        }
-
-        try {
-            TableId parsed = TableId.parse(raw, true);
-            if (!isTableIdCaseSensitive()) {
-                parsed = parsed.toLowercase();
-            }
-            final Set<TableId> matches = readTableNames(
-                    parsed.catalog(), null, parsed.table(), new String[]{ "TABLE" });
-            if (matches.isEmpty()) {
-                return Collections.singletonList(
-                        "Signal data collection '" + raw + "' was not found in the database.");
-            }
-            return SignalDataCollectionChecks.validateShape(raw, readColumns(matches.iterator().next()));
-        }
-        catch (SQLException e) {
-            // Signal table is only exercised when a signal is later inserted; a probe failure here
-            // must not block connector creation. Surface the failure to operators via the log only.
-            LOGGER.warn("Unable to validate signal data collection '{}'; skipping signal-table check", raw, e);
-            return Collections.emptyList();
-        }
+        final Set<TableId> matches = readTableNames(
+                parsed.catalog(), null, parsed.table(), new String[]{ "TABLE" });
+        return matches.isEmpty() ? null : matches.iterator().next();
     }
 
     /**
