+++ debezium-core/src/main/java/io/debezium/jdbc/JdbcConnection.java
@@ -59,6 +59,7 @@ import io.debezium.relational.Attribute;
 import io.debezium.relational.Column;
 import io.debezium.relational.ColumnEditor;
 import io.debezium.relational.RelationalDatabaseConnectorConfig;
+import io.debezium.relational.SignalDataCollectionChecks;
 import io.debezium.relational.Table;
 import io.debezium.relational.TableId;
 import io.debezium.relational.Tables;
@@ -1410,6 +1411,55 @@ public class JdbcConnection implements AutoCloseable {
             }
         }
         return columns;
+    }
+
+    /**
+     * Validates the configured signal data collection (when validation is enabled) against the
+     * documented shape: the table exists and has exactly three columns named {@code id},
+     * {@code type}, {@code data} at positions 0/1/2. Returns an empty list when validation is
+     * disabled, when {@code signal.data.collection} is unset, or when a probe-time
+     * {@link SQLException} is caught — a validator-side failure must never block connector
+     * CREATE/UPDATE because the signal table is only exercised later when a signal is inserted.
+     * <p>
+     * The DB-specific identifier parse + probe is exposed via
+     * {@link #resolveSignalDataCollectionTableId(String)} so connector-specific
+     * subclasses only override that hook.
+     */
+    public List<String> validateSignalDataCollection(CommonConnectorConfig config) {
+        if (!config.isSignalDataCollectionValidationEnabled()) {
+            return Collections.emptyList();
+        }
+        final String raw = config.getSignalingDataCollectionId();
+        if (Strings.isNullOrBlank(raw)) {
+            return Collections.emptyList();
+        }
+
+        try {
+            final TableId resolved = resolveSignalDataCollectionTableId(raw);
+            if (resolved == null) {
+                return Collections.singletonList(
+                        "Signal data collection '" + raw + "' was not found in the database.");
+            }
+            return SignalDataCollectionChecks.validateShape(raw, readColumns(resolved));
+        }
+        catch (SQLException e) {
+            LOGGER.warn("Unable to validate signal data collection '{}'; skipping signal-table check", raw, e);
+            return Collections.emptyList();
+        }
+    }
+
+    /**
+     * Resolve the {@link TableId} for the signal data collection from the raw config string, or
+     * {@code null} if no such table exists. Default implementation parses the value as
+     * {@code schema.table} (or {@code catalog.schema.table}) with schema-first semantics and
+     * probes once via {@link #readTableNames}. Subclasses should override to apply DB-specific
+     * parsing (e.g. catalog-first on MySQL) or identifier folding before probing.
+     */
+    protected TableId resolveSignalDataCollectionTableId(String raw) throws SQLException {
+        final TableId parsed = TableId.parse(raw, false);
+        final Set<TableId> matches = readTableNames(
+                parsed.catalog(), parsed.schema(), parsed.table(), new String[]{ "TABLE" });
+        return matches.isEmpty() ? null : matches.iterator().next();
     }
 
     protected Map<TableId, List<Attribute>> getAttributeDetails(TableId tableId, String tableType) {
